requirements.txt
google-auth-oauthlib
google-api-python-client
requests
python-dotenv
kaggle
diffusers
torch
python-dotenv